refactor: refactor characters api to RTK query
