+++ rs-react-app/src/components/ItemDetails/ItemDetails.tsx
@@ -1,12 +1,6 @@
-import React, { useEffect, useState, useRef } from 'react';
-
-interface ItemDetails {
-  name: string;
-  gender: string;
-  height: string;
-  mass: string;
-  hairColor: string;
-}
+// ItemDetails.tsx
+import React, { useEffect, useRef } from 'react';
+import { useGetItemDetailsQuery } from './../../slices/itemDetailsSlice.ts';
 
 interface ItemDetailsProps {
   url: string;
@@ -19,29 +13,8 @@ const ItemDetails: React.FC<ItemDetailsProps> = ({
   onClose,
   position,
 }) => {
-  const [details, setDetails] = useState<ItemDetails | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
+  const { data: details, error, isLoading } = useGetItemDetailsQuery(url);
   const modalRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const fetchDetails = async () => {
-      setLoading(true);
-      setError(null);
-      try {
-        const response = await fetch(`${url}`);
-        if (!response.ok) throw new Error('Failed to load details');
-        const data = await response.json();
-        setDetails(data);
-      } catch (err) {
-        setError(err instanceof Error ? err.message : 'Error loading details');
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchDetails();
-  }, [url]);
 
   useEffect(() => {
     const handleClickOutside = (event: MouseEvent) => {
@@ -66,10 +39,10 @@ const ItemDetails: React.FC<ItemDetailsProps> = ({
       <button className="close-button" onClick={onClose}>
         ×
       </button>
-      {loading ? (
+      {isLoading ? (
         <div className="loading">Loading...</div>
       ) : error ? (
-        <div className="error">{error}</div>
+        <div className="error">{error.toString()}</div>
       ) : (
         <div>
           <h2>{details?.name}</h2>
@@ -86,7 +59,7 @@ const ItemDetails: React.FC<ItemDetailsProps> = ({
             <strong>Hair Color:</strong> {details?.hairColor}
           </p>
         </div>
-      )}{' '}
+      )}
     </div>
   );
 };
